python/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
pymongo==4.7.3
python-dotenv==1.0.1
